+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - trapFocus: Added correct handling for focusable html tags with tabIndex=-1
 - Flyout: Updated outside click handlers to only trigger when it's active
 - DropdownMenu.SubTrigger: Added all DropdownMenu.Item props support
+- ScrollArea: handled onScroll x, y values
 
 Theming related: